+++ rs/nns/governance/src/gen/ic_nns_governance.pb.v1.rs
@@ -90,236 +90,20 @@ pub struct NeuronStakeTransfer {
     #[prost(uint64, tag = "7")]
     pub memo: u64,
 }
-/// This structure represents a neuron "at rest" in governance system of
-/// the Internet Computer IC.
-#[derive(candid::CandidType, candid::Deserialize, serde::Serialize, comparable::Comparable)]
-#[compare_default]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct Neuron {
-    /// The id of the neuron.
-    ///
-    /// This is stored here temporarily, since its also stored on the map
-    /// that contains neurons.
-    ///
-    /// Initialization uses ids for the following graph. We need neurons
-    /// to come into existence at genesis with pre-chosen ids, so a
-    /// neuron needs to have an id. We could alternatively choose a
-    /// unique naming scheme instead and chose the ids on the
-    /// initialization of the canister.
-    #[prost(message, optional, tag = "1")]
-    pub id: ::core::option::Option<::ic_nns_common::pb::v1::NeuronId>,
-    /// The principal of the ICP ledger account where the locked ICP
-    /// balance resides. This principal is indistinguishable from one
-    /// identifying a public key pair, such that those browsing the ICP
-    /// ledger cannot tell which balances belong to neurons.
-    #[prost(bytes = "vec", tag = "2")]
-    #[serde(with = "serde_bytes")]
-    pub account: ::prost::alloc::vec::Vec<u8>,
-    /// The principal that actually controls the neuron. The principal
-    /// must identify a public key pair, which acts as a “master key”,
-    /// such that the corresponding secret key should be kept very
-    /// secure. The principal may control many neurons.
-    #[prost(message, optional, tag = "3")]
-    pub controller: ::core::option::Option<::ic_base_types::PrincipalId>,
-    /// Keys that can be used to perform actions with limited privileges
-    /// without exposing the secret key corresponding to the principal
-    /// e.g. could be a WebAuthn key.
-    #[prost(message, repeated, tag = "4")]
-    pub hot_keys: ::prost::alloc::vec::Vec<::ic_base_types::PrincipalId>,
-    /// The amount of staked ICP tokens, measured in fractions of 10E-8
-    /// of an ICP.
-    ///
-    /// Cached record of the locked ICP balance on the ICP ledger.
-    ///
-    /// For neuron creation: has to contain some minimum amount. A
-    /// spawned neuron with less stake cannot increase its dissolve
-    /// delay.
-    #[prost(uint64, tag = "5")]
-    pub cached_neuron_stake_e8s: u64,
-    /// The amount of ICP that this neuron has forfeited due to making
-    /// proposals that were subsequently rejected or from using the
-    /// 'manage neurons through proposals' functionality. Must be smaller
-    /// than 'neuron_stake_e8s'. When a neuron is disbursed, these ICP
-    /// will be burned.
-    #[prost(uint64, tag = "6")]
-    pub neuron_fees_e8s: u64,
-    /// When the Neuron was created. A neuron can only vote on proposals
-    /// submitted after its creation date.
-    #[prost(uint64, tag = "7")]
-    pub created_timestamp_seconds: u64,
-    /// The timestamp, in seconds from the Unix epoch, corresponding to
-    /// the time this neuron has started aging. This is either the
-    /// creation time or the last time at which the neuron has stopped
-    /// dissolving.
-    ///
-    /// This value is meaningless when the neuron is dissolving, since a
-    /// dissolving neurons always has age zero. The canonical value of
-    /// this field for a dissolving neuron is `u64::MAX`.
-    #[prost(uint64, tag = "8")]
-    pub aging_since_timestamp_seconds: u64,
-    /// The timestamp, in seconds from the Unix epoch, at which this
-    /// neuron should be spawned and its maturity converted to ICP
-    /// according to <https://wiki.internetcomputer.org/wiki/Maturity_modulation.>
-    #[prost(uint64, optional, tag = "19")]
-    pub spawn_at_timestamp_seconds: ::core::option::Option<u64>,
-    /// Map `Topic` to followees. The key is represented by an integer as
-    /// Protobuf does not support enum keys in maps.
-    #[prost(map = "int32, message", tag = "11")]
-    pub followees: ::std::collections::HashMap<i32, neuron::Followees>,
-    /// Information about how this neuron voted in the recent past. It
-    /// only contains proposals that the neuron voted yes or no on.
-    #[prost(message, repeated, tag = "12")]
-    pub recent_ballots: ::prost::alloc::vec::Vec<BallotInfo>,
-    /// `true` if this neuron has passed KYC, `false` otherwise
-    #[prost(bool, tag = "13")]
-    pub kyc_verified: bool,
-    /// The record of the transfer that was made to create this neuron.
-    #[prost(message, optional, tag = "14")]
-    pub transfer: ::core::option::Option<NeuronStakeTransfer>,
-    /// The accumulated unstaked maturity of the neuron, in "e8s equivalent".
-    ///
-    /// The unit is "e8s equivalent" to insist that, while this quantity is on
-    /// the same scale as ICPs, maturity is not directly convertible to ICPs:
-    /// conversion requires a minting event and the conversion rate is variable.
-    #[prost(uint64, tag = "15")]
-    pub maturity_e8s_equivalent: u64,
-    /// The accumulated staked maturity of the neuron, in "e8s equivalent" (see
-    /// "maturity_e8s_equivalent"). Staked maturity becomes regular maturity once
-    /// the neuron is dissolved.
-    ///
-    /// Contrary to `maturity_e8s_equivalent` this maturity is staked and thus
-    /// locked until the neuron is dissolved and contributes to voting power
-    /// and rewards. Once the neuron is dissolved, this maturity will be "moved"
-    /// to 'maturity_e8s_equivalent' and will be able to be spawned (with maturity
-    /// modulation).
-    #[prost(uint64, optional, tag = "20")]
-    pub staked_maturity_e8s_equivalent: ::core::option::Option<u64>,
-    /// If set and true the maturity rewarded to this neuron for voting will be
-    /// automatically staked and will contribute to the neuron's voting power.
-    #[prost(bool, optional, tag = "21")]
-    pub auto_stake_maturity: ::core::option::Option<bool>,
-    /// Whether this neuron is "Not for profit", making it dissolvable
-    /// by voting.
-    #[prost(bool, tag = "16")]
-    pub not_for_profit: bool,
-    /// If set, this neuron is a member of the Community Fund. This means that when
-    /// a proposal to open an SNS token swap is executed, maturity from this neuron
-    /// will be used to participate in the SNS token swap.
-    #[prost(uint64, optional, tag = "17")]
-    pub joined_community_fund_timestamp_seconds: ::core::option::Option<u64>,
-    /// If set, the neuron belongs to the "known neurons". It has been given a name and maybe a description.
-    #[prost(message, optional, tag = "18")]
-    pub known_neuron_data: ::core::option::Option<KnownNeuronData>,
-    /// The type of the Neuron. See \[NeuronType\] for a description
-    /// of the different states.
-    #[prost(enumeration = "NeuronType", optional, tag = "22")]
-    pub neuron_type: ::core::option::Option<i32>,
-    /// See the Visibility enum.
-    #[prost(enumeration = "Visibility", optional, tag = "23")]
-    pub visibility: ::core::option::Option<i32>,
-    /// The last time that voting power was "refreshed". There are two ways to
-    /// refresh the voting power of a neuron: set following, or vote directly. In
-    /// the future, there will be a dedicated API for refreshing. Note that direct
-    /// voting implies that refresh also occurs when a proposal is created, because
-    /// direct voting is part of proposal creation.
-    ///
-    /// Effect: When this becomes > 6 months ago, the amount of voting power that
-    /// this neuron can exercise decreases linearly down to 0 over the course of 1
-    /// month. After that, following is cleared, except for ManageNeuron proposals.
-    ///
-    /// This will always be populated. If the underlying neuron was never
-    /// refreshed, this will be set to 2024-11-05T00:00:01 UTC (1730764801 seconds
-    /// after the UNIX epoch).
-    #[prost(uint64, optional, tag = "24")]
-    pub voting_power_refreshed_timestamp_seconds: ::core::option::Option<u64>,
-    /// The index of the next entry in the recent_ballots list that will be
-    /// used for the circular buffer. This is used to determine which entry
-    /// to overwrite next.
-    #[prost(uint32, optional, tag = "25")]
-    pub recent_ballots_next_entry_index: ::core::option::Option<u32>,
-    /// The maturity disbursements that are in progress for this neuron.
-    #[prost(message, repeated, tag = "28")]
-    pub maturity_disbursements_in_progress: ::prost::alloc::vec::Vec<MaturityDisbursement>,
-    /// At any time, at most one of `when_dissolved` and
-    /// `dissolve_delay` are specified.
-    ///
-    /// `NotDissolving`. This is represented by `dissolve_delay` being
-    /// set to a non zero value.
-    ///
-    /// `Dissolving`. This is represented by `when_dissolved` being
-    /// set, and this value is in the future.
-    ///
-    /// `Dissolved`. All other states represent the dissolved
-    /// state. That is, (a) `when_dissolved` is set and in the past,
-    /// (b) `dissolve_delay` is set to zero, (c) neither value is set.
-    ///
-    /// Cf. \[Neuron::stop_dissolving\] and \[Neuron::start_dissolving\].
-    #[prost(oneof = "neuron::DissolveState", tags = "9, 10")]
-    pub dissolve_state: ::core::option::Option<neuron::DissolveState>,
-}
-/// Nested message and enum types in `Neuron`.
-pub mod neuron {
-    /// Protobuf representing a list of followees of a neuron for a
-    /// specific topic.
-    #[derive(
-        candid::CandidType,
-        candid::Deserialize,
-        serde::Serialize,
-        comparable::Comparable,
-        Clone,
-        PartialEq,
-        ::prost::Message,
-    )]
-    pub struct Followees {
-        #[prost(message, repeated, tag = "1")]
-        pub followees: ::prost::alloc::vec::Vec<::ic_nns_common::pb::v1::NeuronId>,
-    }
-    /// At any time, at most one of `when_dissolved` and
-    /// `dissolve_delay` are specified.
-    ///
-    /// `NotDissolving`. This is represented by `dissolve_delay` being
-    /// set to a non zero value.
-    ///
-    /// `Dissolving`. This is represented by `when_dissolved` being
-    /// set, and this value is in the future.
-    ///
-    /// `Dissolved`. All other states represent the dissolved
-    /// state. That is, (a) `when_dissolved` is set and in the past,
-    /// (b) `dissolve_delay` is set to zero, (c) neither value is set.
-    ///
-    /// Cf. \[Neuron::stop_dissolving\] and \[Neuron::start_dissolving\].
-    #[derive(
-        candid::CandidType,
-        candid::Deserialize,
-        serde::Serialize,
-        comparable::Comparable,
-        Clone,
-        Copy,
-        PartialEq,
-        ::prost::Oneof,
-    )]
-    pub enum DissolveState {
-        /// When the dissolve timer is running, this stores the timestamp,
-        /// in seconds from the Unix epoch, at which the neuron becomes
-        /// dissolved.
-        ///
-        /// At any time while the neuron is dissolving, the neuron owner
-        /// may pause dissolving, in which case `dissolve_delay_seconds`
-        /// will get assigned to: `when_dissolved_timestamp_seconds -
-        /// <timestamp when the action is taken>`.
-        #[prost(uint64, tag = "9")]
-        WhenDissolvedTimestampSeconds(u64),
-        /// When the dissolve timer is stopped, this stores how much time,
-        /// in seconds, the dissolve timer will be started with. Can be at
-        /// most 8 years.
-        ///
-        /// At any time while in this state, the neuron owner may (re)start
-        /// dissolving, in which case `when_dissolved_timestamp_seconds`
-        /// will get assigned to: `<timestamp when the action is taken> +
-        /// dissolve_delay_seconds`.
-        #[prost(uint64, tag = "10")]
-        DissolveDelaySeconds(u64),
-    }
+/// Protobuf representing a list of followees of a neuron for a
+/// specific topic.
+#[derive(
+    candid::CandidType,
+    candid::Deserialize,
+    serde::Serialize,
+    comparable::Comparable,
+    Clone,
+    PartialEq,
+    ::prost::Message,
+)]
+pub struct Followees {
+    #[prost(message, repeated, tag = "1")]
+    pub followees: ::prost::alloc::vec::Vec<::ic_nns_common::pb::v1::NeuronId>,
 }
 /// Subset of Neuron that has no collections or big fields that might not exist in most neurons, and
 /// the goal is to keep the size of the struct consistent and can be easily stored in a
@@ -594,7 +378,7 @@ pub struct RewardNodeProviders {
 )]
 pub struct SetDefaultFollowees {
     #[prost(map = "int32, message", tag = "1")]
-    pub default_followees: ::std::collections::HashMap<i32, neuron::Followees>,
+    pub default_followees: ::std::collections::HashMap<i32, Followees>,
 }
 /// Obsolete. Superseded by OpenSnsTokenSwap.
 #[derive(
@@ -783,7 +567,7 @@ pub struct ManageNeuron {
     pub neuron_id_or_subaccount: ::core::option::Option<manage_neuron::NeuronIdOrSubaccount>,
     #[prost(
         oneof = "manage_neuron::Command",
-        tags = "2, 3, 4, 5, 6, 7, 8, 9, 10, 13, 14, 15, 16, 17"
+        tags = "2, 3, 4, 5, 6, 7, 8, 9, 10, 13, 14, 15, 16, 17, 18"
     )]
     pub command: ::core::option::Option<manage_neuron::Command>,
 }
@@ -1302,6 +1086,37 @@ pub mod manage_neuron {
         pub to_account_identifier:
             ::core::option::Option<::icp_ledger::protobuf::AccountIdentifier>,
     }
+    #[derive(
+        candid::CandidType,
+        candid::Deserialize,
+        serde::Serialize,
+        comparable::Comparable,
+        Clone,
+        PartialEq,
+        ::prost::Message,
+    )]
+    pub struct SetFollowing {
+        #[prost(message, repeated, tag = "1")]
+        pub topic_following: ::prost::alloc::vec::Vec<set_following::FolloweesForTopic>,
+    }
+    /// Nested message and enum types in `SetFollowing`.
+    pub mod set_following {
+        #[derive(
+            candid::CandidType,
+            candid::Deserialize,
+            serde::Serialize,
+            comparable::Comparable,
+            Clone,
+            PartialEq,
+            ::prost::Message,
+        )]
+        pub struct FolloweesForTopic {
+            #[prost(message, repeated, tag = "1")]
+            pub followees: ::prost::alloc::vec::Vec<::ic_nns_common::pb::v1::NeuronId>,
+            #[prost(enumeration = "super::super::Topic", optional, tag = "2")]
+            pub topic: ::core::option::Option<i32>,
+        }
+    }
     /// The ID of the neuron to manage. This can either be a subaccount or a neuron ID.
     #[derive(
         candid::CandidType,
@@ -1357,6 +1172,8 @@ pub mod manage_neuron {
         RefreshVotingPower(RefreshVotingPower),
         #[prost(message, tag = "17")]
         DisburseMaturity(DisburseMaturity),
+        #[prost(message, tag = "18")]
+        SetFollowing(SetFollowing),
     }
 }
 #[derive(candid::CandidType, candid::Deserialize, serde::Serialize, comparable::Comparable)]
@@ -2917,9 +2734,6 @@ pub mod update_canister_settings {
 #[compare_default]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct Governance {
-    /// Current set of neurons.
-    #[prost(btree_map = "fixed64, message", tag = "1")]
-    pub neurons: ::prost::alloc::collections::BTreeMap<u64, Neuron>,
     /// Proposals.
     #[prost(btree_map = "uint64, message", tag = "2")]
     pub proposals: ::prost::alloc::collections::BTreeMap<u64, ProposalData>,
@@ -2982,7 +2796,7 @@ pub struct Governance {
     ///
     /// Default following can be changed via proposal.
     #[prost(map = "int32, message", tag = "13")]
-    pub default_followees: ::std::collections::HashMap<i32, neuron::Followees>,
+    pub default_followees: ::std::collections::HashMap<i32, Followees>,
     /// The maximum time a proposal of a topic with *short voting period*
     /// is open for voting. If a proposal on a topic with short voting
     /// period has not been decided (adopted or rejected) within this