+++ rs/sns/governance/src/gen/ic_sns_governance.pb.v1.rs
@@ -644,6 +644,95 @@ pub struct RegisterDappCanisters {
     #[prost(message, repeated, tag = "1")]
     pub canister_ids: ::prost::alloc::vec::Vec<::ic_base_types::PrincipalId>,
 }
+#[derive(
+    candid::CandidType,
+    candid::Deserialize,
+    comparable::Comparable,
+    Clone,
+    PartialEq,
+    ::prost::Message,
+)]
+pub struct PreciseValue {
+    #[prost(oneof = "precise_value::PreciseValue", tags = "1, 2, 3, 4, 5, 6, 7")]
+    pub precise_value: ::core::option::Option<precise_value::PreciseValue>,
+}
+/// Nested message and enum types in `PreciseValue`.
+pub mod precise_value {
+    #[derive(
+        candid::CandidType,
+        candid::Deserialize,
+        comparable::Comparable,
+        Clone,
+        PartialEq,
+        ::prost::Oneof,
+    )]
+    pub enum PreciseValue {
+        #[prost(bool, tag = "1")]
+        Bool(bool),
+        #[prost(bytes, tag = "2")]
+        Blob(::prost::alloc::vec::Vec<u8>),
+        #[prost(string, tag = "3")]
+        Text(::prost::alloc::string::String),
+        #[prost(uint64, tag = "4")]
+        Nat(u64),
+        #[prost(int64, tag = "5")]
+        Int(i64),
+        #[prost(message, tag = "6")]
+        Array(super::PreciseArray),
+        #[prost(message, tag = "7")]
+        Map(super::PreciseMap),
+    }
+}
+#[derive(
+    candid::CandidType,
+    candid::Deserialize,
+    comparable::Comparable,
+    Clone,
+    PartialEq,
+    ::prost::Message,
+)]
+pub struct PreciseArray {
+    #[prost(message, repeated, tag = "1")]
+    pub array: ::prost::alloc::vec::Vec<PreciseValue>,
+}
+#[derive(
+    candid::CandidType,
+    candid::Deserialize,
+    comparable::Comparable,
+    Clone,
+    PartialEq,
+    ::prost::Message,
+)]
+pub struct PreciseMap {
+    #[prost(btree_map = "string, message", tag = "1")]
+    pub map: ::prost::alloc::collections::BTreeMap<::prost::alloc::string::String, PreciseValue>,
+}
+#[derive(
+    candid::CandidType,
+    candid::Deserialize,
+    comparable::Comparable,
+    Clone,
+    PartialEq,
+    ::prost::Message,
+)]
+pub struct ExtensionInit {
+    #[prost(message, optional, tag = "1")]
+    pub value: ::core::option::Option<PreciseValue>,
+}
+#[derive(
+    candid::CandidType,
+    candid::Deserialize,
+    comparable::Comparable,
+    Clone,
+    PartialEq,
+    ::prost::Message,
+)]
+pub struct RegisterExtension {
+    #[prost(message, optional, tag = "1")]
+    pub chunked_canister_wasm: ::core::option::Option<ChunkedCanisterWasm>,
+    #[prost(message, optional, tag = "2")]
+    pub extension_init: ::core::option::Option<ExtensionInit>,
+}
 /// A proposal to remove a list of dapps from the SNS and assign them to new controllers
 #[derive(
     candid::CandidType,
@@ -776,7 +865,7 @@ pub struct Proposal {
     /// of this mapping.
     #[prost(
         oneof = "proposal::Action",
-        tags = "4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20"
+        tags = "4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21"
     )]
     pub action: ::core::option::Option<proposal::Action>,
 }
@@ -888,9 +977,14 @@ pub mod proposal {
         AdvanceSnsTargetVersion(super::AdvanceSnsTargetVersion),
         /// Change the mapping from custom proposal types to topics.
         ///
-        /// Id = 16;
+        /// Id = 16.
         #[prost(message, tag = "20")]
         SetTopicsForCustomProposals(super::SetTopicsForCustomProposals),
+        /// Register an SNS extension canister.
+        ///
+        /// Id = 17.
+        #[prost(message, tag = "21")]
+        RegisterExtension(super::RegisterExtension),
     }
 }
 #[derive(candid::CandidType, candid::Deserialize, comparable::Comparable)]
@@ -2242,11 +2336,11 @@ pub struct GetMetricsRequest {
     PartialEq,
     ::prost::Message,
 )]
-pub struct GetMetricsResponse {
-    #[prost(uint64, optional, tag = "1")]
-    pub num_recently_submitted_proposals: ::core::option::Option<u64>,
-    #[prost(uint64, optional, tag = "2")]
-    pub last_ledger_block_timestamp: ::core::option::Option<u64>,
+pub struct Metrics {
+    #[prost(uint64, tag = "1")]
+    pub num_recently_submitted_proposals: u64,
+    #[prost(uint64, tag = "2")]
+    pub last_ledger_block_timestamp: u64,
 }
 /// Request message for 'get_sns_initialization_parameters'
 #[derive(