+++ rs/sns/governance/src/governance.rs
@@ -1,3 +1,5 @@
+use crate::icrc_ledger_helper::ICRCLedgerHelper;
+use crate::pb::v1::Metrics;
 use crate::{
     canister_control::{
         get_canister_id, perform_execute_generic_nervous_system_function_call,
@@ -53,8 +55,8 @@ use crate::{
             DisburseMaturityInProgress, Empty, ExecuteGenericNervousSystemFunction,
             FailStuckUpgradeInProgressRequest, FailStuckUpgradeInProgressResponse,
             GetMaturityModulationRequest, GetMaturityModulationResponse, GetMetadataRequest,
-            GetMetadataResponse, GetMetricsRequest, GetMetricsResponse, GetMode, GetModeResponse,
-            GetNeuron, GetNeuronResponse, GetProposal, GetProposalResponse,
+            GetMetadataResponse, GetMetricsRequest, GetMode, GetModeResponse, GetNeuron,
+            GetNeuronResponse, GetProposal, GetProposalResponse,
             GetSnsInitializationParametersRequest, GetSnsInitializationParametersResponse,
             Governance as GovernanceProto, GovernanceError, ListNervousSystemFunctionsResponse,
             ListNeurons, ListNeuronsResponse, ListProposals, ListProposalsResponse,
@@ -80,6 +82,7 @@ use crate::{
     },
     types::{is_registered_function_id, Environment, HeapGrowthPotential, LedgerUpdateLock, Wasm},
 };
+
 use candid::{Decode, Encode};
 #[cfg(not(target_arch = "wasm32"))]
 use futures::FutureExt;
@@ -2010,12 +2013,36 @@ impl Governance {
 
     pub async fn get_metrics(
         &self,
-        _request: GetMetricsRequest,
-    ) -> Result<GetMetricsResponse, GovernanceError> {
-        Err(GovernanceError {
-            error_type: 42,
-            error_message: "Unimplemented".to_string(),
+        request: GetMetricsRequest,
+    ) -> Result<Metrics, GovernanceError> {
+        let num_recently_submitted_proposals = self.recent_proposals(request.time_window_seconds);
+        let icrc_ledger_helper = ICRCLedgerHelper::with_ledger(self.ledger.as_ref());
+
+        let last_ledger_block_timestamp = icrc_ledger_helper
+            .get_latest_block_timestamp_seconds()
+            .await
+            .map_err(|error_mesage| {
+                GovernanceError::new_with_message(ErrorType::External, error_mesage)
+            })?;
+
+        Ok(Metrics {
+            num_recently_submitted_proposals,
+            last_ledger_block_timestamp,
         })
+    }
+
+    fn recent_proposals(&self, time_window_seconds: u64) -> u64 {
+        self.proto
+            .proposals
+            .values()
+            .rev()
+            .take_while(|proposal| {
+                self.env
+                    .now()
+                    .saturating_sub(proposal.proposal_creation_timestamp_seconds)
+                    <= time_window_seconds
+            })
+            .count() as u64
     }
 
     /// Starts execution of the given proposal in the background.
@@ -2090,6 +2117,10 @@ impl Governance {
                 self.perform_register_dapp_canisters(register_dapp_canisters)
                     .await
             }
+            Action::RegisterExtension(_) => Err(GovernanceError::new_with_message(
+                ErrorType::InvalidProposal,
+                "RegisterExtension proposals are not supported yet.",
+            )),
             Action::DeregisterDappCanisters(deregister_dapp_canisters) => {
                 self.perform_deregister_dapp_canisters(deregister_dapp_canisters)
                     .await
@@ -6219,6 +6250,9 @@ mod proposal_topics_tests;
 
 #[cfg(test)]
 mod test_helpers;
+
+#[cfg(test)]
+mod get_metrics;
 
 #[cfg(feature = "canbench-rs")]
 mod benches;