+++ rs/universal_canister/impl/src/api.rs
@@ -73,7 +73,6 @@ mod ic0 {
         pub fn global_timer_set(timestamp: u64) -> u64;
         pub fn canister_version() -> u64;
 
-        pub fn mint_cycles(amount: u64) -> u64;
         pub fn mint_cycles128(amount_high: u64, amount_low: u64, dst: u32) -> ();
 
         pub fn is_controller(src: u32, size: u32) -> u32;
@@ -432,11 +431,6 @@ pub fn trap_with(message: &str) -> ! {
     unsafe {
         ic0::trap(message.as_ptr() as u32, message.len() as u32);
     }
-}
-
-/// Mint cycles (only works on CMC).
-pub fn mint_cycles(amount: u64) -> u64 {
-    unsafe { ic0::mint_cycles(amount) }
 }
 
 /// Mint cycles (only works on CMC).