+++ packages/icrc-ledger-types/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 0.1.10
+
+### Changed
+
+- Remove unneeded dependency on `ic-cdk`.
+
 ## 0.1.9
 
 ### Added